Split the BridgeProvider into TokensContext
