+++ src/views/bridge-confirmation/bridge-confirmation.view.tsx
@@ -27,6 +27,7 @@ import { approve, isContractAllowedToSpendToken } from "src/adapters/ethereum";
 import BridgeButton from "src/views/bridge-confirmation/components/bridge-button/bridge-button.view";
 import useBridgeConfirmationStyles from "src/views/bridge-confirmation/bridge-confirmation.styles";
 import ApprovalInfo from "src/views/bridge-confirmation/components/approval-info/approval-info.view";
+import { useTokensContext } from "src/contexts/tokens.context";
 
 const BridgeConfirmation: FC = () => {
   const callIfMounted = useCallIfMounted();
@@ -34,7 +35,8 @@ const BridgeConfirmation: FC = () => {
   const navigate = useNavigate();
   const env = useEnvContext();
   const { notifyError } = useErrorContext();
-  const { bridge, tokens } = useBridgeContext();
+  const { bridge } = useBridgeContext();
+  const { tokens } = useTokensContext();
   const { formData, setFormData } = useFormContext();
   const { openSnackbar } = useUIContext();
   const { account, connectedProvider } = useProvidersContext();
